Create markov_chain_an_example
The function buildWordDict takes in the string of text, which was retrieved from the
Internet. It then does some cleaning and formatting, removing quotes, and putting
spaces around other punctuation so it is effectively treated as a separate word. After
this, it builds a two-dimensional dictionary—a “dictionary of dictionaries—that has
the following form:
{word_a : {word_b : 2, word_c : 1, word_d : 1},
word_e : {word_b : 5, word_d : 2},...}
In this example dictionary, “word_a” was found four times, two instances of which
were followed by “word_b,” one instance followed by “word_c,” and one instance fol‐
lowed by “word_d.” “Word_e” was followed seven times, five times by “word_b” and
twice by “word_d.”
If we were to draw a node model of this result, the node representing “word_a” would
have a 50% arrow pointing toward “word_b” (which followed it two out of four
times), a 25% arrow pointing toward “word_c,” and a 25% arrow pointing toward
“word_d.”
Once this dictionary is built up, it can be used as a lookup table to see where to go
next, no matter which word in the text you happen to be on.3 Using the sample dictio‐
nary of dictionaries, we might currently be on “word_e,” which means that we’ll pass
the dictionary {word_b : 5, word_d: 2} to the retrieveRandomWord function. This
function in turn retrieves a random word from the dictionary, weighted by the num‐
ber of times it occurs.
By starting with a random starting word (in this case, the ubiquitous “I”), we can tra‐
verse through the Markov chain easily, generating as many words as we like.